Support non-linear migration plans
In order to reuse the Drifter machinery, we capture the state of the database schema
and thread that through the whole migration process. Drifter's `migrate` function
will order the changes by dependency and then run the migration process from the
very beginning. This is ideal, but we need a way to prevent existing migrations from
being recommitted. This change threads the names of all previous migrations through
the migration process allowing migrate to skip any migrations that are already
done.